pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "soccer-physics-engine"
version = "0.1.0"
description = "Physics-based soccer analytics platform for tactical and load analysis."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kloppy>=3.15",
    "polars>=1.0",
    "pandas>=2.0",
    "pandera>=0.20",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "networkx>=3.2",
    "torch>=2.2",
    "torch-geometric>=2.5",
    "mlflow>=2.10",
    "fastapi>=0.110",
    "pydantic>=2.6",
    "uvicorn>=0.27",
    "matplotlib>=3.8",
    "mplsoccer>=1.2",
    "plotly>=5.18",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
    "mypy>=1.8",
    "httpx>=0.27",
]
pose = [
    "mediapipe>=0.10",
    "opencv-python>=4.9",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "D"]
ignore = ["D203", "D213"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["D"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
packages = ["src"]
